// reservas-web/src/types/Business.ts
export interface Business {
  id: number;
  name: string;
  description?: string;
  category?: string;
  city?: string;
  address?: string;
  phone?: string;
  email?: string;
  imageUrl?: string;
  rating?: number;
  totalResources?: number;
  totalReservations?: number;
  active: boolean;
  ownerId?: number;
}